Painter Can Work with Closure Phase Only
In case there is no field OIVIS and so no differential phase, Painter
will work only with Closure. Modification made for data without OIVIS
field. If OIVIS and so if the field visphi is existing Painter will use
both Phase Closure and visphi to work.

« visphi » can be differential phases computed with a fixed reference
channel (assumption) or estimated phases of complexe visibilities.